Run archiver for yesterday's events for first 5h of day
