+++ images/events-archiver/Dockerfile
@@ -7,4 +7,4 @@ ADD . .
 RUN python3 -m pip install --no-cache-dir -r requirements.txt
 
 
-CMD ["/srv/run.bash"]
+CMD ["/srv/run.py"]
