+++ images/events-archiver/indexer.py
@@ -23,6 +23,61 @@ HERE = os.path.dirname(os.path.abspath(__file__))
 # Static files that should be uploaded by indexer
 STATIC_FILES = glob(os.path.join(HERE, 'static', 'bootstrap-4.1.3.min.css'))
 
+
+def index_events(project, bucket, debug=False, dry_run=False):
+    storage_client = storage.Client()
+    bucket = storage.Bucket(storage_client, bucket)
+
+    blobs = bucket.list_blobs(prefix='events-')
+
+    archives = []
+
+    with open(os.path.join(HERE, 'index.html')) as f:
+        html_template = jinja2.Template(f.read())
+
+    for blob in blobs:
+        archives.append({
+            'name': blob.name,
+            'date': blob.metadata['Events-Date'],
+            'count': blob.metadata['Events-Count']
+        })
+    
+    with tempfile.TemporaryFile(mode='w+') as htmlfile, \
+         tempfile.TemporaryFile(mode='w+') as jsonlfile:
+
+        html_index = html_template.render(
+            archives=sorted(archives, key=lambda archive: archive['date'], reverse=True),
+            generated_time=datetime.utcnow().isoformat() + 'Z'
+        )
+
+        if debug:
+            print(html_index)
+
+        htmlfile.write(html_index)
+
+        for archive in archives:
+            jsonlfile.write(json.dumps(archive) + '\n')
+
+        htmlfile.seek(0)
+        jsonlfile.seek(0)
+
+        if not dry_run:
+            html_blob = bucket.blob('index.html')
+            html_blob.upload_from_file(htmlfile, content_type='text/html')
+            print('Uploaded index.html')
+            bucket.blob('index.jsonl').upload_from_file(jsonlfile)
+            print('Uploaded index.jsonl')
+
+
+    # Upload static assets
+    for static_file in STATIC_FILES:
+        blob_name = os.path.relpath(static_file, HERE)
+        static_blob = bucket.blob(blob_name)
+        mimetype, _ = mimetypes.guess_type(static_file)
+        static_blob.upload_from_filename(static_file, content_type=mimetype)
+        print(f'Uploaded static file {blob_name}')
+        
+
 def main():
     argparser = argparse.ArgumentParser()
     argparser.add_argument(
@@ -51,58 +106,8 @@ def main():
 
     args = argparser.parse_args()
 
+    index_events(args.project, args.bucket, args.debug, args.dry_run)
 
-    storage_client = storage.Client()
-    bucket = storage.Bucket(storage_client, args.bucket)
-
-    blobs = bucket.list_blobs(prefix='events-')
-
-    archives = []
-
-    with open(os.path.join(HERE, 'index.html')) as f:
-        html_template = jinja2.Template(f.read())
-
-    for blob in blobs:
-        archives.append({
-            'name': blob.name,
-            'date': blob.metadata['Events-Date'],
-            'count': blob.metadata['Events-Count']
-        })
-    
-    with tempfile.TemporaryFile(mode='w+') as htmlfile, \
-         tempfile.TemporaryFile(mode='w+') as jsonlfile:
-
-        html_index = html_template.render(
-            archives=sorted(archives, key=lambda archive: archive['date'], reverse=True),
-            generated_time=datetime.utcnow().isoformat() + 'Z'
-        )
-
-        if args.debug:
-            print(html_index)
-
-        htmlfile.write(html_index)
-
-        for archive in archives:
-            jsonlfile.write(json.dumps(archive) + '\n')
-
-        htmlfile.seek(0)
-        jsonlfile.seek(0)
-
-        if not args.dry_run:
-            html_blob = bucket.blob('index.html')
-            html_blob.upload_from_file(htmlfile, content_type='text/html')
-            bucket.blob('index.jsonl').upload_from_file(jsonlfile)
-
-
-    print(STATIC_FILES, file=sys.stderr)
-
-    # Upload static assets
-    for static_file in STATIC_FILES:
-        blob_name = os.path.relpath(static_file, HERE)
-        static_blob = bucket.blob(blob_name)
-        mimetype, _ = mimetypes.guess_type(static_file)
-        static_blob.upload_from_filename(static_file, content_type=mimetype)
-        
 
 if __name__ == '__main__':
     main()
